@@ -98,42 +98,27 @@
 }
 
 @section Scripts {
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.22.2/moment.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/fullcalendar/3.9.0/fullcalendar.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/fullcalendar/3.9.0/locale/es.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-datetimepicker/4.17.47/js/bootstrap-datetimepicker.min.js"></script>
+    <script src="https://code.jquery.com/ui/1.12.1/jquery-ui.min.js"></script>
+    <!-- SignalR scripts -->
+    <script src="https://cdn.jsdelivr.net/npm/@@aspnet/signalr@1.0.2/dist/browser/signalr.min.js" asp-fallback-src="~/js/min/signalr/signalr.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/msgpack5@4.2.0/dist/msgpack5.min.js" asp-fallback-src="~/js/min/signalr/msgpack5.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/@@aspnet/signalr-protocol-msgpack@1.0.2/dist/browser/signalr-protocol-msgpack.min.js" asp-fallback-src="~/js/min/signalr/signalr-protocol-msgpack.min.js"></script>
+
+    <link href="https://cdnjs.cloudflare.com/ajax/libs/fullcalendar/3.9.0/fullcalendar.min.css" rel="stylesheet" />
+    <link href="https://cdnjs.cloudflare.com/ajax/libs/fullcalendar/3.9.0/fullcalendar.print.css" rel="stylesheet" media="print" />
+    <link href="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-datetimepicker/4.17.47/css/bootstrap-datetimepicker.min.css" rel="stylesheet" />
+    <link href="https://cdnjs.cloudflare.com/ajax/libs/jqueryui/1.12.1/jquery-ui.min.css" rel="stylesheet" />
+
     <environment Names="Development">
-        <script src="~/lib/moment/moment.js" asp-append-version="true"></script>
-        <script src="~/lib/fullcalendar/dist/fullcalendar.js" asp-append-version="true"></script>
-        <script src="~/lib/fullcalendar/dist/locale/es.js" asp-append-version="true"></script>
-        <script src="~/lib/eonasdan-bootstrap-datetimepicker/build/js/bootstrap-datetimepicker.min.js" asp-append-version="true"></script>
-        <script src="~/lib/jquery-ui/jquery-ui.js"></script>
-        <!-- SignalR scripts -->
-        <script src="~/js/signalr/signalr.js"></script>
-        <script src="~/js/signalr/msgpack5.js"></script>
-        <script src="~/js/signalr/signalr-protocol-msgpack.js"></script>
-
         <script src="~/js/pacientes/pacientes.citas.js" asp-append-version="true"></script>
-
-        <link href="~/lib/fullcalendar/dist/fullcalendar.css" rel="stylesheet" />
-        <link href="~/lib/fullcalendar/dist/fullcalendar.print.css" rel="stylesheet" media="print" />
-        <link href="~/lib/eonasdan-bootstrap-datetimepicker/build/css/bootstrap-datetimepicker.css" rel="stylesheet" />
-        <link href="~/lib/jquery-ui/themes/base/jquery-ui.css" rel="stylesheet" />
     </environment>
 
     <environment Names="Production,Staging">
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.22.2/moment.min.js"></script>
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/fullcalendar/3.9.0/fullcalendar.min.js"></script>
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/fullcalendar/3.9.0/locale/es.js"></script>
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-datetimepicker/4.17.47/js/bootstrap-datetimepicker.min.js"></script>
-        <script src="https://code.jquery.com/ui/1.12.1/jquery-ui.min.js"></script>
-        <!-- SignalR scripts -->
-        <script src="https://cdn.jsdelivr.net/npm/@@aspnet/signalr@1.0.2/dist/browser/signalr.min.js" asp-fallback-src="~/js/min/signalr/signalr.min.js"></script>
-        <script src="https://cdn.jsdelivr.net/npm/msgpack5@4.2.0/dist/msgpack5.min.js" asp-fallback-src="~/js/min/signalr/msgpack5.min.js"></script>
-        <script src=" https://cdn.jsdelivr.net/npm/@@aspnet/signalr-protocol-msgpack@1.0.2/dist/browser/signalr-protocol-msgpack.min.js" asp-fallback-src="~/js/min/signalr/signalr-protocol-msgpack.min.js"></script>
-
         <script src="~/js/min/pacientes/pacientes.citas.min.js" asp-append-version="true"></script>
-
-        <link href="https://cdnjs.cloudflare.com/ajax/libs/fullcalendar/3.9.0/fullcalendar.min.css" rel="stylesheet" />
-        <link href="https://cdnjs.cloudflare.com/ajax/libs/fullcalendar/3.9.0/fullcalendar.print.css" rel="stylesheet" media="print" />
-        <link href="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-datetimepicker/4.17.47/css/bootstrap-datetimepicker.min.css" rel="stylesheet" />
-        <link href="~/lib/jquery-ui/themes/base/jquery-ui.min.css" rel="stylesheet" />
     </environment>
 
     <script type="text/javascript">
