@@ -1,9 +1,4 @@
-﻿@using System.Collections.Generic
-@using System.Linq
-@using Microsoft.AspNetCore.Http
-@using Microsoft.AspNetCore.Http.Authentication
-@using Akari_Net.Core.Areas.Usuarios.Models.Entities
-@using Web.Areas.Usuarios.Data
+﻿@using Web.Areas.Usuarios.Data
 @model LoginViewModel
 @inject SignInManager<ApplicationUser> SignInManager
 
